fix: align hierarchy action icons
